Changed the API a bit
Now, each lecture has ONE quiz, and each quiz belongs to ONE lecture. Each quiz has one or many questions(I've added the properties of the quiz that you originally made to the questions, like question text or correctAnswerIndex), you can submit the score of the quiz by sending the number of correct answers(so the frontend can either show a percentage or X / Y). Fixed up the favorites. API is in the correct order.
Now we need to add a way to submit quizzes and question for the lectures through the website