Changed some things around
the favorites were previously for the teachers and lectures, it was supposed to be for the teachers and courses.
Now the API sends the score when receiving the lectures.
Changed the relation between the course and the lecture, i remember it gave me an error before but i forgot where(let's hope it doesn't happen again, lol)